GroupLevel_FREQNESS_NetworkEstimation.m
This updated version of the main function introduces support for group-level analysis.
It now accepts both 2D (channels × time) and 3D (channels × time × participants) data matrices as input. When 3D input is provided, participants are stored along the last dimension in the fields of the output structure.
The resulting structure can then be passed to all other second-order functions, which will automatically account for the participant dimension when present.