GroupLevel_FREQNESS_Visualizer.m
This updated version of the visualizer function introduces support for group-level analysis.
It now accepts input structures from either single participants or groups of participants, where participants are stored by default in the last dimension of each field in the structure.
Network landscapes and spatial activation patterns are generated for all individual participants as well as for the group-level grand average.

NOTE on spatial activation patterns:
The 3D visualization displayed in MATLAB is intended only as a rudimentary preview. For accurate inspection of topographies, we recommend using the NIFTI files produced in the output folder (e.g., with FSLeyes or similar software).